REFACTOR SPLIT FILES
